Symbols for each column
I crosschecked the symbols with a purple pen and it seems like kind match column for all columns except column E which contains both kind E and D. I changed from all symbols to what I hope are the correct symbols for each column.

There was an issue with svg-control-descriptions. Symbol "bend" was assigned to wrong kind in lang.json.